+++ web-wallet/src/lib/components/__tests__/Send.spec.js
@@ -7,7 +7,6 @@ import {
 	vi
 } from "vitest";
 import { cleanup, fireEvent, render } from "@testing-library/svelte";
-
 import { deductLuxFeeFrom } from "$lib/contracts";
 import { createCurrencyFormatter } from "$lib/dusk/currency";
 import { getAsHTMLElement } from "$lib/dusk/test-helpers";
@@ -23,9 +22,11 @@ describe("Send", () => {
 		formatter,
 		gasSettings: {
 			gasLimit: 20000000,
+			gasPrice: 1
+		},
+		gasLimits: {
 			gasLimitLower: 10000000,
-			gasLimitUpper: 1000000000,
-			gasPrice: 1,
+			gasLimitUpper: 2900000000,
 			gasPriceLower: 1
 		},
 		spendable: 1000,